Test reorganized into categories, and size reset
There are getting to be too many of them, I might have to split them
soon, which will be easier if they are broken into sections.

Reorganizing tests revealed a dependency on size being undefined at
start of a test, which requires size to be set to undefined on stop.
This particular test used to run before master was ever run, so
master.size was always undefined. The test is arguably wrong, but
I prefer master state to be reset on stop, so I made the test pass.