Remove loadOptions(), its no longer used
It was useful so that the many direct users of strong-cluster-control
could have consistent command line argument usage, but since
strong-supervisor's run command has come in, that's no longer a common
usage pattern. It also depended on rc, which at the time made it
consistent with loopback, but now seems a down side. It has non-standard
environment variable names, non-standard argument processing, and made
it harder to add features to strong-supervisor... the only current user
of the API.