emit custom 'fork' events with s-c-c metadata
These are the same as those emitted by cluster, but the worker instance
passed along is the strong-cluster-control augmented version.

Also emit a 'fork' event on the workers themselves so that a specific
worker can be correlated with the cluster.fork() call that created it
and the s-c-c metadata will have been set on the worker object.