Fix busy-loop on resize after disconnect
Resize was assuming that there would always be a worker to disconnect,
and busy-looping with nextTick if there was not.

The state of having a cluster.worker that is disconnected is temporary,
it will be removed once its exit status has been received. However, that
status cannot be received if node is looping in next tick.

On v0.10, this isn't noticed very often because nextTick is limited in
how deep it will recurse before returning to the event loop. The only
symptom would be a brief spike of ~100% CPU usage until the worker
exits.

On v0.12+, nextTick recursion is unlimited, so this can block
indefinitely. Changing to setImmediate solves the indefinite problem,
but still briefly busy loops.

The fix is for stopOne() to wait for a worker that has already been
disconnected to exit before continuing.