Restart should start new then stop old
It was doing the opposite. This would allow the total number of workers
to be as much a 2 less than the desired size. On a 2 core box, this
could mean no workers at all. Even with more cores, it isn't desireable,
there should always be the configured number of workers available to
service requests.