status: preserve the cluster size during restart
cluster restart involves changing the cluster size as workers are
started and stopped. This was reflected in the status().master.setSize
output, but that's confusing to consumers of that data (such as arc)
because the size is changing all the time. Tracing is racy, it relies on
the assumption that the size doesn't change at critical times, if so, it
can get stuck. This reduces the prevalence of those races.